+++ src/main/java/com/ssh/dartserver/domain/team/application/TeamService.java
@@ -11,6 +11,9 @@ import com.ssh.dartserver.domain.team.infra.TeamRepository;
 import com.ssh.dartserver.domain.team.infra.TeamUserRepository;
 import com.ssh.dartserver.domain.team.domain.TeamSearchCondition;
 import com.ssh.dartserver.domain.team.presentation.response.*;
+import com.ssh.dartserver.domain.team.util.RegionMapper;
+import com.ssh.dartserver.domain.team.util.TeamMapper;
+import com.ssh.dartserver.domain.team.util.TeamViewCountNotificationUtil;
 import com.ssh.dartserver.domain.user.domain.User;
 import com.ssh.dartserver.domain.user.domain.profilequestions.ProfileQuestions;
 import com.ssh.dartserver.domain.user.presentation.v1.response.ProfileQuestionResponse;
